Refactor: Props 수정
